+++ RPOSC-LogicAnalyser/src/MeasuredData.hpp
@@ -8,7 +8,7 @@ class MeasuredData: public SContainer {
     public:
         MeasuredData(std::string _name, int _size, std::string _def_value);
         void Update();
-        void addData(std::string acqChannel, std::vector<double> data);
+        void addData(std::string acqChannel, std::vector<float> data);
         void resetData();
         std::vector<std::string> measuredData;
 };
